requirements.txt
flask
torch
transformers
pybind11
pdfminer.six